+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Fixes
 - Fixed date input in /rfudrophistory add not working.
+- Made hud not be hidden while on inventory
 
 ### Back-end
 - Added an event manager for fishing sessions and made fishing HUD elements update reactively.